Avoid reflection in CryptoConfigShims
https://github.com/dotnet/SignService/issues/266

The previous code attempts to invoke an internal method in .NET Core as
if it's non-generic, but in recent framework versions it _is_ generic.
This commit replaces the call with a non-reflective call on a public
method to avoid similar backwards-breaking changes in the future.

Note: There are some special cases handled by the internal method
System.Security.Cryptography.Xml.CryptoHelpers.CreateFromName that
are not handled by the replacement method, CryptoConfig.CreateFromName.
However, the only names used in CryptoConfigShims are for
RSAPKCS1SignatureFormatter and RSAPKCS1SignatureDeformatter, neither of
which appear to be in the list of special cases. So for the purposes of
CryptoConfigShims, I think the new behaviour is equivalent.