Enable user protected keys in certificate store
Sign needs to be able to use user-protected keys, such as keys stored
in a USB eToken device. These generally require user interaction
when the key is used and signing will fail when dialogs are blocked.

User interaction requires the 'Silent' flag to not be set in
CertificateStoreService.GetRsaAsync method. The rest of this is
plumbing to make a user option available to this code.

The new option is --user-protected-key/-upk to turn off the silent
flag. The flag name corresponds to the enum value
CspProviderFlags.UseUserProtectedKey in
System.Security.Cryptography. I preferred this to not-silent or
-silent false (with a true default).

I'm sure there is a process for translation, but I do not know if this
is done with the initial pull request.